requirements-win.txt
future
numpy<=1.14.6
scipy<=1.1.0
networkx<=2.2
requests<=2.20.0
pandas<=0.23.4
Matplotlib<3.0.0
seaborn<=0.9.0
pyteomics<=3.5.1
tqdm<=4.28.1
biopython<=1.72
python-qt5